requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended
Flask-Cors
scikit-learn==1.3.0  
numpy==1.23.5 
python-dotenv
tensorflow
pandas
seaborn
imblearn
xgboost
